#98 Tornando o profile_picture um componente global
